requirements.txt
#streamlit>=1.28.0
#langchain>=0.1.0
#langchain-groq>=0.1.0
#langchain-openai>=0.1.0
#requests>=2.25.0
#feedparser>=6.0.0
#beautifulsoup4>=4.9.0
#yake>=0.4.8
#python-dotenv>=0.19.0

streamlit>=1.28.0
langchain>=0.0.309,<0.1.0
langchain-groq==0.1.6
langchain-openai>=0.1.0
requests>=2.25.0
feedparser>=6.0.0
beautifulsoup4>=4.9.0
yake>=0.4.8
python-dotenv>=0.19.0
